Minor tweaks

+++ src/main.rs
@@ -8,19 +8,17 @@ use clap::Clap;
     author = "Benjamin Falk <benjamin.falk@yahoo.com>"
 )]
 struct Opts {
-    /// number of sides to the dice, should be
-    /// between 1 and 255
+    /// number of sides to the dice
     #[clap(short, long, default_value = "20")]
     sides: u8,
 
-    /// number of times you want to roll the dice
+    /// number of times to roll the dice
     #[clap(short, long, default_value = "1")]
     times: usize,
 }
 
 fn main() {
     let opts: Opts = Opts::parse();
-
     let mut rng = rand::thread_rng();
 
     for roll in 1..(opts.times + 1) {
@@ -29,7 +27,7 @@ fn main() {
         println!(
             "Roll #{roll:->5} : {value:>3}",
             roll = roll,
-            value = (num * opts.sides as f32 + 1.) as usize
+            value = (num * opts.sides as f32 + 1.) as u8
         );
     }
 }
